reorganized components structures
