+++ frontend/budgetai/src/components/Authentication/utils/AuthContext.js
