+++ frontend/budgetai/src/components/Authentication/utils/GoogleSignUp.js
@@ -1,6 +1,6 @@
-import "../css/Login.css"
+import "../../../css/Login.css"
 import { getAuth, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
-import app from "../firebase/firebaseConfig"; 
+import app from "../../../firebase/firebaseConfig"; 
 import {useNavigate} from "react-router-dom";
 import { useCookies } from "react-cookie";
