+++ frontend/budgetai/src/components/Authentication/utils/ProtectedRoute.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../components/AuthContext";
+import { useAuth } from "./AuthContext";
 import checkAuth from "./checkAuth"; // Import the checkAuth utility
 
 const ProtectedRoute = ({ children }) => {
