+++ frontend/budgetai/src/components/Authentication/utils/checkAuth.js
